Change types for capital FinancingSummary back to BigDecimal

+++ src/main/java/com/stripe/model/capital/FinancingSummary.java
@@ -106,7 +106,7 @@ public class FinancingSummary extends ApiResource {
      * milliseconds since unix epoch.
      */
     @SerializedName("advance_paid_out_at")
-    Long advancePaidOutAt;
+    BigDecimal advancePaidOutAt;
 
     /** Currency that the financing offer is transacted in. For example, {@code usd}. */
     @SerializedName("currency")
@@ -133,7 +133,7 @@ public class FinancingSummary extends ApiResource {
      * epoch.
      */
     @SerializedName("repayments_begin_at")
-    Long repaymentsBeginAt;
+    BigDecimal repaymentsBeginAt;
 
     /** Per-transaction rate at which Stripe will withhold funds to repay the financing. */
     @SerializedName("withhold_rate")
@@ -150,7 +150,7 @@ public class FinancingSummary extends ApiResource {
        * seconds since unix epoch.
        */
       @SerializedName("due_at")
-      Long dueAt;
+      BigDecimal dueAt;
 
       /** The amount that has already been paid in the current repayment interval. */
       @SerializedName("paid_amount")
